requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
transformers>=4.30.0
torch>=2.0.0
toml>=0.10.2
pillow-avif-plugin>=1.3.1
numpy>=1.24.0
pathlib>=1.0.1